+++ pyscript/lib/ast/ast_literal_eval.py
@@ -21,11 +21,20 @@ get_identifier = {
     'nanj': pys_builtins.nanj
 }.get
 
+def _error(exception, position):
+    return type(exception)(
+        '{} (ln {}, col {})'.format(
+            exception,
+            position.start_line,
+            position.start_column
+        )
+    )
+
 def visit(node: PysNode) -> Any:
     return get_visitor(node.__class__)(node)
 
 def visit_unknown_node(node: PysNode) -> None:
-    raise ValueError(f"invalid node: {type(node).__name__}")
+    raise _error(ValueError(f"invalid node: {type(node).__name__}"), node.position)
 
 def visit_NumberNode(node: PysNumberNode) -> int | float | complex:
     return node.value.value
@@ -41,7 +50,7 @@ def visit_IdentifierNode(node: PysIdentifierNode) -> Any:
     name = node.name.value
     value = get_identifier(name)
     if value is None:
-        raise ValueError(f"invalid identifier: {name}")
+        raise _error(ValueError(f"invalid identifier: {name}"), node.position)
     return value
 
 def visit_DictionaryNode(node: PysDictionaryNode) -> dict:
@@ -60,19 +69,19 @@ def visit_CallNode(node: PysCallNode) -> set:
     target = node.target
     if isinstance(target, PysIdentifierNode) and target.name.value == 'set' and not node.arguments:
         return set()
-    raise ValueError("invalid call node except for 'set()'")
+    raise _error(ValueError("invalid call node except for 'set()'"), node.position)
 
 def visit_UnaryOperatorNode(node: PysUnaryOperatorNode) -> Any:
     operand = node.operand.type
     if is_arithmetic(operand):
         return GET_UNARY_FUNCTION(operand)(visit(node.value))
-    raise ValueError(f"invalid unary operator node: {REVERSE_TOKENS[operand]}")
+    raise _error(ValueError(f"invalid unary operator node: {REVERSE_TOKENS[operand]}"), node.position)
 
 def visit_BinaryOperatorNode(node: PysBinaryOperatorNode) -> Any:
     operand = node.operand.type
     if is_arithmetic(operand):
         return GET_BINARY_FUNCTION(operand)(visit(node.left), visit(node.right))
-    raise ValueError(f"invalid binary operator node: {REVERSE_TOKENS[operand]}")
+    raise _error(ValueError(f"invalid binary operator node: {REVERSE_TOKENS[operand]}"), node.position)
 
 def visit_EllipsisNode(node: PysEllipsisNode) -> EllipsisType:
     return ...