+++ test.py
@@ -29,3 +29,26 @@ def pyscript_doc():
         input=f'<pre>{pyscript.pys_highlight(source.strip())}</pre>',
         encoding='utf-8'
     )
+
+def update_snippets():
+    import json
+
+    with open('./highlight/vscode/snippets/pyscript.json') as file:
+        data = json.load(file)
+
+    with open('./highlight/acode/src/snippets.js', 'w') as file:
+        content = []
+
+        for name, snippet in data.items():
+            prefix = snippet['prefix']
+            body = snippet['body']
+            if isinstance(body, list):
+                body = '\n'.join(body)
+            body = pyscript.core.utils.string.indent(body, 1, '\t')
+            # type = name.split(' - ')[0].lower()
+            content.append(f'snippet {prefix}\n{body}')
+
+        content = '\n\n'.join(content) + '\n'
+        file.write(f'export const snippets = {content!r};')
+
+update_snippets()